Refactor protocol UI registrations into service extensions
